Interface map editor working
