+++ CMakeLists.txt
@@ -20,13 +20,15 @@ include_directories(${SOURCE_DIR})
 set(SOURCE_FILES src/main.cpp
         src/Interface.cpp
         src/Map_editor_handler.cpp
+        src/Button_entry.cpp
+        stb_image.cpp
         src/Menu.cpp)
 
 #src/Container_map.cpp
 #src/glad.c src/Map_parser.cpp src/Parameters_label.cpp src/Scene.cpp
 #src/Map_editor_handler.cpp
 
-add_executable(${PROJECT_NAME} ${SOURCE_DIR}/main.cpp stb_image.cpp)
+add_executable(${PROJECT_NAME} ${SOURCE_FILES})
 #add_executable(Object Objects.cpp)
 
 #add_executable(Shader shader_s.h)
@@ -36,7 +38,7 @@ add_executable(${PROJECT_NAME} ${SOURCE_DIR}/main.cpp stb_image.cpp)
 
 SET_TARGET_PROPERTIES(Quiz_Figure PROPERTIES LINKER_LANGUAGE C)
 
-SET(CMAKE_C_FLAGS "-pthread  -Wall -Werror  -fPIC ")
+SET(CMAKE_C_FLAGS " -pthread  -Wall -Werror  -fPIC ")
 # SET(CMAKE_CXX_FLAGS "-lglfw3 -lGL -lX11 -lpthread -lXrandr -lXi -ldl")
 # SET(CMAKE_C_FLAGS "-Wall -Werror -Wpedantic -fPIC")
 
@@ -68,11 +70,7 @@ file(GLOB sources
 
 ###########################
 
-add_library(map_editor_lib STATIC src/Interface.cpp include/Interface.h
-        src/Map_editor_handler.cpp include/Map_editor_handler.h
-        src/Menu.cpp include/Menu.h)
 
-target_link_libraries(${PROJECT_NAME} PRIVATE map_editor_lib)
 
 ###########################
 
@@ -92,3 +90,9 @@ TARGET_LINK_LIBRARIES(${PROJECT_NAME} PUBLIC glad ${GLFW_STATIC_LIBRARIES})
 
 ####################
 
+add_library(map_editor_lib STATIC src/Interface.cpp include/Interface.h
+        src/Map_editor_handler.cpp include/Map_editor_handler.h
+        src/Button_entry.cpp include/Button_entry.h
+        src/Menu.cpp include/Menu.h )
+
+target_link_libraries(${PROJECT_NAME} PRIVATE map_editor_lib)
